requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-filter==2.2.0
django-hashers-passlib==0.4
django-rest-auth==0.9.5
djangorestframework==3.10.3
idna==2.8
Markdown==3.1.1
oauthlib==3.1.0
passlib==1.7.1
Pygments==2.4.2
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
RestAuthCommon==0.7.0
six==1.13.0
South==1.0.2
sqlparse==0.3.0
urllib3==1.25.7
